EDSC-914: process request parameters before forwarding them to CMR in dataaccess controller.
